requirements.txt
numpy==1.26.4
ollama==0.5.4
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-ollama==0.3.7
chromadb==1.1.0
docling==2.53.0
docling-core==2.48.1
docling-ibm-models==3.9.1
docling-parse==4.5.0
pytesseract==0.3.13
easyocr==1.7.2
opencv-python-headless==4.11.0.86
fastapi==0.117.1
uvicorn==0.36.0
pandas==2.3.2
pillow==11.3.0
python-multipart==0.0.20
pydantic==2.11.7
python-dotenv==1.1.1
requests==2.32.5
pytest==8.3.5